accounting-agent/backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==1.10.13
pytest==8.3.5
sentence-transformers==2.2.2
transformers==4.30.2
huggingface_hub==0.14.1
chromadb==0.4.24
scikit-learn==1.6.1
